+++ winsane/winsane_option.cpp
@@ -5,7 +5,7 @@
  *                 | |/ |/ / / /_/ /___/ / /_/ / / / /  __/
  *                 |__/|__/_/\__,_//____/\__,_/_/ /_/\___/
  *
- * Copyright (C) 2012 - 2013, Marc Hoersken, <info@marc-hoersken.de>
+ * Copyright (C) 2012 - 2014, Marc Hoersken, <info@marc-hoersken.de>
  *
  * This software is licensed as described in the file COPYING, which
  * you should have received as part of this software distribution.
@@ -442,7 +442,7 @@ PVOID WINSANE_Option::GetValue(_In_ SANE_Word value_type, _In_ SANE_Word value_s
 	SANE_Status status;
 	SANE_Word info, pointer;
 	SANE_String resource;
-	DWORD written;
+	LONG written;
 	PBYTE buf;
 
 	buf = new BYTE[value_size];
@@ -487,7 +487,7 @@ PVOID WINSANE_Option::SetValue(_In_ SANE_Word value_type, _In_ SANE_Word value_s
 	SANE_Status status;
 	SANE_Word info, pointer;
 	SANE_String resource;
-	DWORD written;
+	LONG written;
 	PBYTE buf;
 
 	buf = new BYTE[value_size];
